add embeddable project pledge button
The button is designed to be hot-linked from a project's website.
This is the simple png version.

This implementation uses the Diagrams library to fill in the number of
patrons on the button template. It should be simple to adapt if the art
changes. The png file is embedded into the program at compile time, and
the png should only get parsed once and then be cached.

I have not benchmarked yesod serving up the buttons, but I expect it to be
pretty fast.