Fix SD-337
We didn't use the renderTime widget in the table of sent invites. Due
to this it wasn't possible to mouse over the time to see the absolute
time. This commit adjusts the tables of sent invites to use renderTime,
which in turn enables the user to mouse over for the absolute time.