Make license notice less visually aggressive
The old license notice when contributing a chance to a wiki page was
very visually aggressive; it was in a red box with a big font. The
license notice shouldn't alarm users, but instead inform them. For this
reason this commit makes the notice less obtrusive using a small font
and calmer colors (grey).