updated formula illustrations
Ideally, this would be smaller and the text not part of the image. Text
should instead be html, etc.

Also, this illustration would work far better animated.

This also provides a possible foundation for some ways we could
illustrate actual graphs of matching. I.e. for live data on the site, we
could show users an image like this for each of their pledges, all the
way down to tiny dots and/or consolidated in some way once things are at
much higher numbers. But this could provide a real visualization of the
effect of someone's pledge.

Anyway, lots of room for improvement. This is just a better and more
accurate presentation going along with the updated formula as well.