total revision of notifications page templates
This is clunky and redundant because I didn't figure out how to remove
the redundancy in the calling of the two nearly-identical pages for
archived or non-archived notifications. SD-614 and the thread there
discusses these issues

This is entirely an improvement over the nothing we had before, but it
should be cleaned up, of course.

Given the situation, I used some awkward in-line CSS and some other
things that aren't ideal but are still good enough to go forward for
now.