Pass pgUser and sudoUser as a command line argument.
Currently sdm assumes postgres is using peer authentication and that
postgres is also a postgresql superuser.  On OpenBSD, postgres is a
postgresql superuser, but the system user who owns the database files.
On Freebsd, both the systemuser and the sudouser are called pgsql.

This allows both the postgres superuser and the sudo target user to be
passed as commandline arguments, so it should work on
OpenBSD/FreeBSD/DragonflyBSD, and probably OSX.

This additionally replaces "sed -i" with perl since the -i flag is
GNU specific.

I mostly used the ReaderT because I'd never actually used one before,
I'm not sure it's that big of a win over thread the params around, so I can
switch it to that style if it's prefered.