SETUP* -> GUIDE, less redundant GUIDE/README
At this point, various system details remain incomplete, the
organization isn't perfect, some of this should probably be moved to the
Wiki, but this is an improvement.

The main outstanding issue is updating the Vagrant stuff to use Stack so
that everything uses Stack consistently. I didn't touch that.

Also, the Nix stuff in the GUIDE appendix probably should be cleaned up.
Will we have people use Stack even with Nix / NixOS? None of that is
clear to me…