Add libravatar support
Until now, if there is no avatar URL, no other source is used to display an
avatar. With this commit, if there is no avatar URL but the user does have a
verified email address, the email address is passed through libravatar to get
an avatar URL from it.

Things to consider modifying:

- Fallback URL (currently none is specified)
- Size (currently it's the default, 80)
- HTTPS (currently an HTTP URL is generated)