back-slashes not needed to close html in-line
I previously guessed that since hamlet auto-closes html tags, I needed
to *not* start a line with a tag if I wanted to manually close it before
the end of the line (such as before a comma). Well, it turns out that I
can just have the manual closing tag, and it's accepted regardless, so
no need to avoid the plain starting tag at the beginning of a line.