fixes SD-614 query for archived notifications
The notification archive stuff is quite a mess, but this at least makes
it consistent with other things and more consolidated.

Because it was too complex for me to solve, I basically combined
redundant code into the same hamlet file but didn't eliminate all the
redundancy. While the separate /archived route is gone, I didn't
eliminate the proxy route for the form submission.

With this change, everything works consistently with the notifications
UI, there is no longer a separate archived page to get out of sync.

Eventually, the whole redundancy in the backend and basic data structure
should be addressed, but everything is working here and is improved
clarity and easier to maintain than before.