clean stub page for project browse /p
Left some hamlet commented so we can use it to add back in project logos
once we decide how that will work.

Removed all the ticket count and discussion and even patron count stuff,
any of that we want in the long run will need to have that code updated
or retrieved.

This is a far more acceptable stub page than the broken thing we had.