improved error
If I mark a user eligible, then hit back, I can submit again and it
won't work, but the error that came back wasn't appropriate. This is
just a for-now better solution. The way these errors look and are
handled should be improved, but it's not a priority. This is an odd case
anyway.