crowdmatch: Use (SqlPersistT env) monad to allow combining queries
Before this commit, crowdmatch actions run database queries separately
inside them like a black box. That means in practice many applications
of `runDB`, one per query, so a transaction per query. This is an
unnecessary waste, and doesn't gain much (or anything at all, as this
commit's changes my demonstrate).

So this commit adds `SqlPersistT` on top of the crowdmatch env monad, so
that the actual running of SQL queries is controlled from outside
crowdmatch itself, and allows a single `runDB` to handle a batch of
persistent actions.

Both Snowdrift and the crowdmatch test suite have been updated to make
use of this new capability, which also happens to make the code more
readable (e.g. much much less occurences of `runDB`).

Signed-off-by: Bryan Richter <b@chreekat.net>