Lookup-ability Done Simpler [Main Commit of Push]
Of all the commits in this push, this one most directly pertains to accomplishing our objective:
enabling us to lookup the definition of any Haskell function (or other declared entity) we use,
in a way simpler than last.
The other commits are either in response to this one (to make other things match and work with
the new ways of things) or are in response to pre-existing conversation.

Our newfound lookup-ability comes from tag.sh, formerly unpack.sh.
This script downloads and unpacks Haskell dependencies into a git-ignored subdirectory. Then,
it generates a "tags" file for all Haskell entities in our repo's code, as well as those in its
dependencies. A text editor may then load this file and let you lookup any Haskell entity we
use.

We simplified unpack.sh into tag.sh:

  In the old way of doing things, sdb.hs was not compiled, but ran as a script. This was done
  with a crunchbang on line one, which told the kernel to use Stack as the interpreter, and a
  special comment to Stack on line two that specified sdb's dependencies.

  Because the command unpack.sh used to gather the list of dependencies did not pick up on those
  of sdb, instead only looking at stack.yaml and Cabal files, extra code was added to unpack.sh
  to get sdb's dependencies from a fake package we made for this purpose. And we would have
  had to manually maintain that fake package as other Haskell scripts got added.

  To simplify unpack.sh into tag.sh, we agreed to convert the existing Haskell script, sdb.hs,
  into an ordinary Haskell source file in a Cabal package. This elimated the "scripts" fake
  package, as well as the extra code in unpack.sh. We also agreed to never go the Haskell script
  route again, now aware of the complication it brings.

At first, we had considered sdb to be a development tool. But upon reflection, it seems to be
more of an administrator's tool for managing a Snowdrift database server.
So, we made subdir "admin-tools" and moved sdb.hs into there. Also there, we created
admin-tools.cabal to make the package that now contains sdb.hs. This new package was of course
added to stack.yaml: necessary to make tag.sh work in a simple fashion.

The crunchbang and dependency lines were removed from sdb.hs, and its executable flag was also
removed. It's help text was also updated to be more beginner-friendly and make implementation
an aside, putting abstraction first. And it was added that the "clean" command stops the server
process.

We deleted definition-lookup-aid/scripts.cabal and [ditto]/stack.yaml, the files which
comprised the fake scripts package.

According to suggestion, a dev-tools subdir was made. The now tags.sh was moved there.
The .gitignore file was edited accordingly, to ignore our subdirectory for unpacking
dependencies, which goes wherever the script goes.

Also, tags.sh does the tagging now. The old script only downloaded and unpacked.