Non-Working State: Locating project root for .postgres-work
Now that sdb is compiled, we are not in control of where the executable file goes:
Stack is. This means we cannot determine the project root relative to its location.
And yet we want to put .postgres-work at the root (at time of writing). So how do we
do it?

Well, now we have sdb.sh. It serves as a wrapper to the binary, passing the desired
directory path as an environmental variable. It figures this out from its location.
There is also a fallback to sdb.sh that builds the binary locally, should build.sh
have problems.

Please read the comments in sdb.sh for more info.

I propagated this change throughout the project: things now run/document `sdb.sh [arg]`
instead of `stack exec sdb [arg]`. I hadn't realized before that sdb needed to know
where the project root is and that it was doing so based off where the core executable
file is.

I also added some comments, escaping quotes and made a couple changes to .gitignore.