Broken-State: Polish, Fixed Bash double-dot, Fixed data type issue
* sdb.sh was using the Linux double-dot directory link to get to the project root.
This caused an issue. It now just uses dirname twice. I don't know why it didn't
work, but I appeased it.
NB: Had a problem with Stack saying that sdb is in two places when I try
'stack exec ghc ./sdb.hs ./sdb`. Maybe this is from it remembering the double-dot?

* sdb.hs had a data type problem from my previous work. It is now fixed.

* Polished some doc/code comments.

* Hid the fallback sdb binary so it wouldn't be an eye-sore (prefixed name with a dot).