Bookmarking Working State
I was using `stack build` illegally in build.sh. I passed the command two package targets, Snowdrift and admin-tools. However, the command strictly
takes zero or one targets. At time of writing, if you give two, it does the first and ignores the second. So it never even tried to build
admin-tools, which contains sdb. So it didn't even try to compile sdb.

That is now fixed. The build script now uses two `stack build` lines.

And for good measure, I had the script cd to the directory containing it, in absolute form, first thing after a little argument processing.