Display login failure above, not below, warning about auth system
It seems Yesod stores messages in the session in reverse order, i.e. it
prepends each new message to the beginning of the string. So when
traversing messages in default layout, we get them in order reverse to
the actual order we added them.

To make the login failur appear first, addDanger second in the code,
after the warning rather than before.