Gave heads-up comments about Makefile plan
Bryan has decided that sdb.hs will be replaced with
a Makefile, so I gave a comment at the top of sdb.hs
saying so. This is to save the reader time, encouraging
them to study code elsewhere in the repo.

As another heads-up, a note was left at the top of
build.sh, informing developers that it will change
or go away altogether, due to said Makefile.

Note: I do plan to implement the Makefile solution.