First wave of tweaks: special chars & filesystem shenanigans + other fixes
Fixed logic with the `stop` rule in `.sdc`.

Made environment varibles in `config` hold absolute paths.

Fixed a mistake: sdsh does *not* accept params, so "$@" was removed.

Made `.pgStart` to deal with special characters and documented it.

Made hidden Makefiles non-executable to somewhat enforce going through the
interface.

Only prefaced commands with "./" if need be - not with use of `make -f`.

Added breif note that editors may recognize tab-char-indented files in local
README.