Share horizontal padding evenly on medium layout navbar elements
This affects the clickable area of elements on the navbar (so moving the mouse
cursor slightly left of the P of "Projects" now keeps "Projects" highlighted
rather than highlighting "How it works"). The navbar's appearance shouldn't
change.

For the purpose of adding navbar drop-downs, this also makes it easier to align
drop-down elements with their navbar label.