Enable grid visual overlays during development
Remove comment

  /// Skip this section for production.

since we're (still) unconditionally including rules for the visual grid.
These rules apply to <.mk__grid>, which, under development conditions,
is now added to every page that uses the new design.

I've also changed the appearance of the grid, as I thought that the
default looked too subtle.