Rearrange "running locally" content in BUILD.md
Information about change-detecting and re-running was spread out (the note about static files was below the original info). This moves them together, and then shuffles the other sections so that makes sense. It also puts the sections in order from less to more involved; info about logging in should be higher than connecting to the database.