Split GIT-BASICS out of CONTRIBUTING.md
CONTRIBUTING.md has a lot of content oriented at beginners. This makes
it difficult to skim and not very useful to experienced developers.

The git section was a particularly egregious example, with almost
nothing specific to snowdrift.coop. Clearing it out makes room for
future documentation that would help more with onboarding.