Add about status box
Very inelegant, a total temporary job just to get this status page live
while I continue to work on the prototype - I've just stuffed everything
I need from the prototype to make this work into _status-box-temp.silius

It's not exactly the same as on the prototype, because the current
styles change the base font size at the "tiny" media query width, so the
sizes of some things are different. But it looks decent enough, not
worth spending more time fooling around with it.

This has partially validated my planned approach for migrating the
prototype pages -- converting the html to hamlet was super fast and then
the vast majority of the work was fiddling with the sass.. so if we can
import the sass wholesale to avoid fiddling, it should be much less
effort, totally do-able