Add status flag
Even hackier than the status box: A ton of stuff is inlined into
_status-flag-tamp.sass to avoid potentially conflicting with any other
styles around the site (since the navbar is included on every page).

Also a few places were converted from rem to px due to the current main
styles changing the root font size at respond("xs") while the prototype
does not do that.

It's also not quite horizontally aligned, but mucking with sass with 60s
build times is painful, so it'll have to do.