Show tenths of a cent. Fixes #113
Previously the math as displayed to a viewer of the website didn't add
up, even though it was correct if you considered tenths of a cent. So
we'll show them.

- Modify the Project data type to store income in DonationUnits
- Add a new rendering function so DonationUnits are rendered with tenths
  of a cent.

I decided to leave the rendering of Cents in place, this way it's clear
when something is internally calculated as a rounded total (won't show
fractions of a cent) vs a precise number (will show).

The code isn't the cleanest or the most idiomatic, and showing 3 decimal
places for dollars is a bit awkward, but it's part of the frontend which
we're planning on rewriting in a few days anyway, so it's fine.