Module-ize handlers for faster incremental builds
For some reason, all of the pages using "v2" styles compile very slowly.

Before: Changing any route that was declared in Handler.hs required
rebuilding the whole file, which took 48 seconds. Ugh.

After: Each slow-to-build route is in its own file, the slowest of which
takes 14 seconds to build. This doesn't speed up the initial build, but
it means when iterating, changing a page will "only" take 8-14 seconds.

This is as fast as it'll go for now— for reasons unknown to me, even a
simple hamlet file with no splices takes 8s to build (or maybe this is
the splices in the wrapper).

Note: All times were measured on my machine, "by hand", which means
there's a 1 to 2 second margin of error.