Add make-payments variant to charge team members
For now, that means charging only Stephen (myself). It's hard coded to
my user ID in the production database. This approach obviously doesn't
scale at all, but it doesn't need to, since this is just for testing
purposes, before we start charging everyone with the regular
run-payments program. Also, we plan to migrate off of Haskell, so the
maintainability of this code base is not a concern, either.