Use more expressive data type for team-only charge
Having a specific data type for TeamOnly vs AllPatrons makes it clearer
what's going on here.

Also, remove an incorrect comment. It was added before implementation of
team-only was started, describing the approach we were going to take,
but then we had trouble with that approach and used a different one, so
the comment was just wrong.