Add executable that handles passwords
As part of a migration strategy off Haskell, as outlined at
https://community.snowdrift.coop/t/plan-for-moving-off-haskell-part-1-overview/1826/1

Password must be base64 encoded because:

Node's exec function doesn't escape parameters, so we need to protect
against command injection (ex: a password like '; sudo do_evil').  We
*could* try to escape their input, but a bug in that code might allow
bypassing the escaping, so I feel most secure taking the nuclear option
and just base64 encoding the user's password, which I know is
un-escape-able, because base64 doesn't include shell special characters.
The SnowdriftPassword parses from base64 on the other side.  The only
downside to this approach is less efficiency, so it won't scale as well
and we're more vulnerable to [D]DOS.  But this is a temporary solution
while we are low-traffic, and I am more concerned with being hacked than
taken offline, so I like this trade-off.